backend/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
black==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
click==8.1.3
constantly==15.1.0
cryptography==38.0.4
daphne==4.0.0
Django==4.1.3
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.4

packaging==21.3
pathspec==0.10.2
platformdirs==2.5.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygit2==1.12.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pytz==2022.6
redis==4.5.4
requests==2.28.1
service-identity==21.1.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.13
